backend/requirements.txt
fastapi
uvicorn
pytesseract
google-genai
supabase
python-multipart
pdf2image
pillow
pydantic
python-dotenv
